Report Counters using the same behavior as StatsD
The current implementation is not really compatible with what the
reference StatsD implementation does. StatsD aggregates the increase in
the counter during the flushing interval, and at flush, reports the
current value as a differential and resets the counter.

The implementation in Collectd instead never resets the counter, and
always reports the *absolute* value to RRD as a DERIVE, assuming it will
perform the differential itself. This behavior is rather surprising,
particularly when graphing with tools that expect to be
Graphite-compatible.

This patch implements the right behavior by reporting counters as a
"gauge" of the differential (i.e. resetting the counter between flush
intervals). This mimics StatsD's behavior, and even the same behavior
that Collectd performs when reporting the "count" row in a histogram --
after all, the count row in the histogram should have the same semantics
as a counter metric. This was previously inconsistent in this
implementation.